WAR: remove warnings for .NET compiler

+++ Source/cmCMakeMinimumRequired.cxx
@@ -30,7 +30,7 @@ bool cmCMakeMinimumRequired::InitialPass(std::vector<std::string> const& args)
     m_Makefile->AddDefinition("CMAKE_MINIMUM_REQUIRED_VERSION", args[1].c_str());
     }
   float version = float(m_Makefile->GetMajorVersion());
-  version += (float(m_Makefile->GetMinorVersion()) * .1);
+  version += (float(m_Makefile->GetMinorVersion()) * (float).1);
   float reqVersion = 0;
   sscanf(args[1].c_str(), "%f", &reqVersion);
   if(reqVersion > version)
